backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
langchain==0.2.1
langchain-community==0.2.1
langchain-groq==0.1.3
langchain-huggingface==0.0.1
sentence-transformers==2.7.0
torch==2.2.2
faiss-cpu==1.8.0
youtube-transcript-api==0.6.2
beautifulsoup4==4.12.3
requests==2.32.2
python-dotenv==1.0.1
fastembed>=0.3.0
langchain-text-splitters>=0.2.0
# openai==1.30.1 (Removed)
